requirements.txt
# Core dependencies for 3D functionality
numpy>=1.26.4
torch>=2.5.1
transformers==4.51.0.dev0  # Using development version
sentence-transformers==3.2.1
sentencepiece>=0.1.99
accelerate>=0.26.1
requests>=2.28.0

# Bluesky API dependencies
atproto==0.0.7

# 3D and AI dependencies
anthropic==0.38.0  # Using currently installed version
mpx-genai-sdk>=0.1.0
bitsandbytes>=0.41.3
optimum>=1.14.1
auto-gptq>=0.4.2  # Only need one entry for auto-gptq
einops>=0.7.0
confluent-kafka
kafka-python>=2.0.2  # Added for Redpanda client